Move EC2 to base, clean up unused code, move APIs to shared, add delete Domain on delete env
