--- frontend/src/modules/Environments/services/createMLStudioDomain.js
@@ -1,19 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const createMLStudioDomain = (input) => ({
-  variables: {
-    input
-  },
-  mutation: gql`
-    mutation createMLStudioDomain($input: NewStudioDomainInput) {
-      createMLStudioDomain(input: $input) {
-        sagemakerStudioUri
-        environmentUri
-        label
-        vpcType
-        vpcId
-        subnetIds
-      }
-    }
-  `
-});
