--- frontend/src/modules/Environments/services/deleteEnvironmentMLStudioDomain.js
@@ -1,12 +1,0 @@
-import { gql } from 'apollo-boost';
-
-export const deleteEnvironmentMLStudioDomain = ({ environmentUri }) => ({
-  variables: {
-    environmentUri
-  },
-  mutation: gql`
-    mutation deleteEnvironmentMLStudioDomain($environmentUri: String!) {
-      deleteEnvironmentMLStudioDomain(environmentUri: $environmentUri)
-    }
-  `
-});
