+++ frontend/src/modules/Environments/views/EnvironmentView.js
@@ -34,7 +34,7 @@ import {
   useSettings
 } from 'design';
 import { SET_ERROR, useDispatch } from 'globalErrors';
-import { useClient } from 'services';
+import { deleteEnvironmentMLStudioDomain, useClient } from 'services';
 import { archiveEnvironment, getEnvironment } from '../services';
 import { KeyValueTagList, Stack, StackStatus } from 'modules/Shared';
 import {
@@ -111,6 +111,16 @@ const EnvironmentView = () => {
       })
     );
     if (!response.errors) {
+      if (isModuleEnabled(ModuleNames.MLSTUDIO)) {
+        const response2 = await client.mutate(
+          deleteEnvironmentMLStudioDomain({
+            environmentUri: env.environmentUri
+          })
+        );
+        if (response2.errors) {
+          dispatch({ type: SET_ERROR, error: response.errors[0].message });
+        }
+      }
       handleArchiveObjectModalClose();
       enqueueSnackbar('Environment deleted', {
         anchorOrigin: {
